S3 Native Presigned URLs
Adding pre-signed URLs implementation using s3 native functionality
so that alternative (non AWS) deployments can be achieved

Medplum is nodejs server + postgres + redis plus s3, with a CDN in front.
But you can deploy all of that not just on AWS but on other clouds.
Lots of clouds offer postgres/redis/s3-compatible offerings, notably GCP.
With this small change we allow the usage of native pre-signed s3 functionality
so that it can be deployed in many more arrangements than just AWS