Client-side filtering in requestSchema to avoid loading  extra SDs
Until the v4 fix of `eq` in `_filter` using exact match lands.

Previously, if a StructureDefinition exists without a snapshot element
and has a name that begins with another resourceType, it was
unintentially loaded and indexed by `requestSchema`. The indexing
throws an error "No snapshot defined for StructureDefinition <name>"
that bubbles up leading to unexpected errors, e.g. in @medplum/app.