Fix flakey SearchPopupMenu test (#6077)
* Avoid flaky tests

Opening and closing a Mantine menu repeatedly as these tests
were previously doing was causing tests to fail occasionally. Just
keeping the menu open instead of toggling it open and closed
repeatedly has consistently avoided the flakiness for me locally.

This feels like a reasonable fix since we don't really want to test
the ability of a Mantine menu to open and close rapidly in particular
here.

* Remove unneeded casts