Ensure index name uniqueness (#6640)
* Remove incorrect composite primary key declaration

* Don't allow index name collisions

Previously, if index names were computed to be the same,
it could lead to strange behavior when attempting to reconcile
schema drift. e.g. If two indexes were assigned the name
"HumanName_family_idx" as was previously was happening,
since we include `IF NOT EXISTS` in our `CREATE INDEX ...` queries,
the create index query would be a noop since one of the two indexes
already exists.