Exclude `undefined` from `arrayify` result type when possible
When calling `arrayify` with a value whose type is known not to include
`undefined`, we shouldn't introduce an `undefined` into the return type.

This allows us to remove some type casts that seem like they were
intended to strip that `undefined` back out of the result type.

In practice, some of these type casts were also not fully descriptive:
for example, we were casting `TypedValueWithPath[]` into `TypedValue[]`,
meaning that attempting to access the `path` attribute would have
resulted in an inappropriate type error.

Signed-off-by: Noah Silas <noah@medplum.com>