Fix color scheme segmented control highlight
When using the "pop" family of transitions, `<FloatingIndicator>`
components (such as the one used under the hood of a
`<SegmentedControl>`) behave awkwardly - it appears that they compute
position information based on the element size, and the "pop"
transitions animate that by changing a `scale` transform.

This has been reported in Mantine and they do not plan on addressing
this at the library level:
https://github.com/mantinedev/mantine/issues/3330

We have a couple options here, but the most expedient one is to change
our transition style to something that does not animate `scale`. I've
opted for "fade-down" which still looks good but doesn't trigger this
bug.

Fade-down transition:
https://github.com/mantinedev/mantine/blob/8.3.7/packages/%40mantine/core/src/components/Transition/transitions.ts#L52-L56

Signed-off-by: Noah Silas <noah@medplum.com>