Only apply defaultPatientAccessPolicy when userType is "Patient" (#7825)
* Tidy up SCIM user create missing type information

In [#4134](https://github.com/medplum/medplum/pull/4134) the
`getScimUserResourceType` was updated to return "Practitioner" as a
default value when no other type is provided, seemingly to allow Okta to
provision practitioners without supplying this field.

- Removed a check for `undefined` that can now never fire, removing some
  potential for confusion
- Updated the test that had been marked as `skip` to reflect the
  expected behavior

Signed-off-by: Noah Silas <noah@medplum.com>

* Add explicit type check for createScimUser requester

In [#2837](https://github.com/medplum/medplum/pull/2837) this switched
from an untyped `res.locals.user` to the typed `ctx.login.user`. This
detected an issue: `createScimUser` only accepts a `Reference<User>`,
but the logged in caller is a `Reference<User | Bot |
ClientApplication>`.

The narrower type is required to eventually comply with the type
definition of [ProjectMembership.invitedBy](https://github.com/medplum/medplum/blob/v5.0.2/packages/fhirtypes/dist/ProjectMembership.d.ts#L124).

Instead of leaving a potentially unsafe type cast here, let's make an
explicit assertion about the user type and throw if we don't get what we
expect.

Signed-off-by: Noah Silas <noah@medplum.com>

* Only apply defaultPatientAccessPolicy when userType is "Patient"

When creating a user of type "Practitioner", you should not need to have
a `defaultPatientAccessPolicy` defined, and if it is defined it should
not be applied to the new user.

When creating users of type "Patient" we do continue to enforce this
policy exists and automatically attach it.

Github-Issue: #7761
Signed-off-by: Noah Silas <noah@medplum.com>

* [autofix.ci] apply automated fixes

---------

Signed-off-by: Noah Silas <noah@medplum.com>
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>