Add DarkMode toggle to react storybook
The `storybook-addon-mantine` package suggests using the
`storybook-dark-mode` addon to set up this toggle,  but this package
seems currently pinned to Storybook@8.x, while we are using 9.x.

Vueless UI has forked the library to `@vueless/storybook-dark-mode` and
released versions that support Storybook 9.x and 10.x. They left a note
on the issue for v9 support on the original package:
https://github.com/hipstersmoothie/storybook-dark-mode/issues/295#issuecomment-3084260327

This fork seems to be slowly picking up steam:
https://npmtrends.com/@vueless/storybook-dark-mode-vs-storybook-dark-mode

The 9.x series still has a bit of funkiness that looks to have been
pulled over from the upstream branch: specifically, we're now pulling in
`memoizerific` as a dev dependency, despite it not being used anywhere.
This has already been removed from @vueless/storybook-dark-mode in the
v10 supporting releases.

I briefly looked at updating Storybook to v10, but the Mantine addon
doesn't yet support it:
https://github.com/josiahayres/storybook-addon-mantine/issues/21

Signed-off-by: Noah Silas <noah@medplum.com>