Validate invited users don't exist at different scope
There is potential for confusion when someone has two User records that
are both tied to the same email address and connected to the same
project.

While this is a valid state to be in, we want to ensure that folks are
reaching it intentionally. To this end, we add a new validation:

- When creating a project-scoped user, ensure that there does not
  already exist a membership connected to a server-scoped user with the
  same email address

- When creating a server-scoped user, ensure that there does not already
  exist a membership connected to a project-scoped user with the same
  email address

Github-Issue: #7572
Signed-off-by: Noah Silas <noah@medplum.com>