Resolve `consistent-return` linting errors
The parent commit triggered linter errors when we changed from `return
Promise.resolve();` to `return;` in some async functions. These should
be identical, but the linter now spots that we had a branch where we are
returning a promise as being inconsistent.

This is a false inconsistency, as in an async function a promise is
always returned. But in the interest of making both Sonar and the linter
happy, we can change from `return` to `await` on the remaining branches.

Signed-off-by: Noah Silas <noah@medplum.com>