Add removeAvailability helper function
Another building block in $find operations; given intervals of
available time and intervals of unavailable time, return a new set of
available intervals that have removed all the unavailable time.

Implementation note: By making the assumption that the input intervals
have already been "normalized" such that there are no overlapping
intervals in either the "available" set or "blocked" set, and the
intervals in each set are sorted, we can do a single pass through each
array to compute this.

Signed-off-by: Noah Silas <noah@medplum.com>