Fail if download worker loses a read-write race
The impact of this should be pretty minimal, but between when we load
the latest version of a resource and when we write the patch back to the
DB there was a window where another writer could collide.

Explicitly passing the current version ID in an ifMatch option will fail
if we lost this race. The download job will then get retried.

Signed-off-by: Noah Silas <noah@medplum.com>