Fix autodownload race (#7781)
* Remove extraneous guard in crawler visitors

The type system tells us that these values always are truthy (either a
`TypedValueWithPath` or an `Array<TypedValueWithPath>`). We can simplify
these visitors by removing these guard clauses that never do anything.

Signed-off-by: Noah Silas <noah@medplum.com>

* Exclude `undefined` from `arrayify` result type when possible

When calling `arrayify` with a value whose type is known not to include
`undefined`, we shouldn't introduce an `undefined` into the return type.

This allows us to remove some type casts that seem like they were
intended to strip that `undefined` back out of the result type.

In practice, some of these type casts were also not fully descriptive:
for example, we were casting `TypedValueWithPath[]` into `TypedValue[]`,
meaning that attempting to access the `path` attribute would have
resulted in an inappropriate type error.

Signed-off-by: Noah Silas <noah@medplum.com>

* Prevent race condition in Autodownloader

When adding a resource with multiple attachments subject to the
Autodownloader, each attachment enqueues a Download job. These jobs can
run in parallel, pulling the remote asset into Storage and then trying
to update the reference to the asset in the resource.

There are two flavors of race happening here:

- A given worker would read the Resource early in its lifetime, import
  the file into storage, and then try to update and write back the
  entire resource. This could cause it to overwrite other updates
  already applied by the related workers.

  This has been resolved by switching from `updateResource` to a more
  highly targeted `patchResource` operation. Further, we refresh the
  resource before generating the patch so that we minimize the chance
  that we are patching over a mutated copy.

- When a worker is committing a change, the autodownload scanner sees
  the update and runs again. Unprocessed files are spotted and matched
  again, enqueuing duplicate jobs for them.

  To prevent this I have added a check to only enqueue the
  autodownloader job for a URL when it is in a field that was added or
  changed from the previous version.

Github-Issue: https://github.com/medplum/medplum/issues/7651
Signed-off-by: Noah Silas <noah@medplum.com>

* Document shortcoming and cost tradeoff in mutation detection

Signed-off-by: Noah Silas <noah@medplum.com>

* Bail out on autodownload when original URL no longer found

Signed-off-by: Noah Silas <noah@medplum.com>

* Move pathToJSONPointer into core Typeschema crawler implementation

This is a broadly useful tool suitable for paths emitted from this
crawler specifically (since it does not support any of the more complex
Fhirpath specification). Colocating it helps make that relationship
explicit.

Signed-off-by: Noah Silas <noah@medplum.com>

* [autofix.ci] apply automated fixes

* Fail if download worker loses a read-write race

The impact of this should be pretty minimal, but between when we load
the latest version of a resource and when we write the patch back to the
DB there was a window where another writer could collide.

Explicitly passing the current version ID in an ifMatch option will fail
if we lost this race. The download job will then get retried.

Signed-off-by: Noah Silas <noah@medplum.com>

---------

Signed-off-by: Noah Silas <noah@medplum.com>
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Signed-off-by: Rahul Agarwal <rahul@medplum.com>