Explicitly use npm@11.7.0 in workflows
NPM v11.6.2 has a bug when dealing with attributes in the dependency
graph that was causing it to label more nodes as "peer" than it ought
to. This is fixed in v11.6.3 via https://github.com/npm/cli/pull/8645.

For now, the latest NodeJS v24 tagged release is still bundling NPM
v11.6.2; the problem will eventually go away as Node's `main` branch has
already bumped the bundled `npm` to v11.7.0.

It would be nice if the `setup-node` action we depend on would let us
specify an explicit package manager version to use, but that is still a
pending feature request (https://github.com/actions/setup-node/issues/529).

In the meantime, we can keep our lockfile in good order by explicitly
adding a step to our action that upgrades NPM to a patched version.

I've also rebuilt the lockfile by running `npm install
--strict-peer-deps --prefer-dedupe` under npm v11.7.0, which removes
the extra "peer" attributes that were introduced in commit e365fa405.

Signed-off-by: Noah Silas <noah@medplum.com>