Cut off slot generation once we have filled a page
This is a performance optimization - if someone requests available slots
across a large window and many service types, we don't want to generate
a list with thousands of entries only to return a page of 20.

We now break out of iteration early once the page is full by introducing
a `flatMapMax()` array utility.

Signed-off-by: Noah Silas <noah@medplum.com>