Remove `useRef` during render warnings
I'm going to be updating this component, and so I wanted to start from a
clean warning-free slate.

It looks like these refs were introduced to gain access to the latest
value of the state at the time that we are mutating the state. We can
use the updater callback argument to setState, which passes the current
value in.

This removes warnings that look like this:

> Error: Cannot access refs during render
>
> React refs are values that are not needed for rendering. Refs should
> only be accessed outside of render, such as in event handlers or
> effects. Accessing a ref value (the `current` property) during render
> can cause your component not to update as expected
> (https://react.dev/reference/react/useRef).

Signed-off-by: Noah Silas <noah@medplum.com>