Add `repeat.timeOfDay` support to TimingInput
For upcoming Scheduling work, users will specify availability as
starting at these times.

This was slightly tricky because it is an array of strings that the
viewer can mutate. To support the viewer dropping an element from the
middle of the array without messing up our input elements, we want to
give each array entry a unique key. We do this by decorating the input
array with incrementing integers, and then undecorating them when we
pass the value back into the real onChange handler.

Signed-off-by: Noah Silas <noah@medplum.com>