Improve TimingInput for scheduling fields (#8094)
* Remove `useRef` during render warnings

I'm going to be updating this component, and so I wanted to start from a
clean warning-free slate.

It looks like these refs were introduced to gain access to the latest
value of the state at the time that we are mutating the state. We can
use the updater callback argument to setState, which passes the current
value in.

This removes warnings that look like this:

> Error: Cannot access refs during render
>
> React refs are values that are not needed for rendering. Refs should
> only be accessed outside of render, such as in event handlers or
> effects. Accessing a ref value (the `current` property) during render
> can cause your component not to update as expected
> (https://react.dev/reference/react/useRef).

Signed-off-by: Noah Silas <noah@medplum.com>

* Teach TimingInput to retain daysOfWeek display on reopening

Making this a controlled component means that opening the modal to edit
a Timing with this dayOfWeek field already populated will not require
the viewer to re-populate the values.

Signed-off-by: Noah Silas <noah@medplum.com>

* Display Timing.repeat.timeOfDay independently of timezone

A `timeOfDay` is represented with a string like "08:00", which should
represent "8 AM" without being tied to a specific timezone.

When the `formatTime` function was being used to represent this, it was
including the viewer's timezone offset in the calculations. In my case
(PST timezone, which is offset UTC-8) this would result in the displayed
time being offset by 8 hours.

I didn't want to introduce a potentially breaking change to the
`formatTime` function, so I've introduced a new `formatWallTime`
function that should stay timezone agnostic.

Signed-off-by: Noah Silas <noah@medplum.com>

* Add `repeat.timeOfDay` support to TimingInput

For upcoming Scheduling work, users will specify availability as
starting at these times.

This was slightly tricky because it is an array of strings that the
viewer can mutate. To support the viewer dropping an element from the
middle of the array without messing up our input elements, we want to
give each array entry a unique key. We do this by decorating the input
array with incrementing integers, and then undecorating them when we
pass the value back into the real onChange handler.

Signed-off-by: Noah Silas <noah@medplum.com>

---------

Signed-off-by: Noah Silas <noah@medplum.com>