Remove AuthenticatedRequestContext.system() (#7783)
* Remove AuthenticatedRequestContext.system()

We generally are moving in the direction of Repository
as the single source of truth for resource/project visibility
and reachability; especially as we look more towards a sharded
future. The incomplete/fake authState value being used in
the system AuthenticatedRequestContext was problematic in that
regard.

Signed-off-by: Matt Long <ml@medplum.com>

* Refactor findTerminologyResource to use repo parameter

Rather than reading from the autenticated request context, pass
the repo directly as a parameter.

Signed-off-by: Matt Long <ml@medplum.com>

* Workers don't require context

Signed-off-by: Matt Long <ml@medplum.com>

* Add tests

Signed-off-by: Matt Long <ml@medplum.com>

* Address TODO

Signed-off-by: Matt Long <ml@medplum.com>

---------

Signed-off-by: Matt Long <ml@medplum.com>