Remove `act` wrappers around `user-event` functions
These functions handle the `act` wrapper internally. Wrapping them with
an `act` call ourselves generates this confusing warning:

> The current testing environment is not configured to support act(...)

My understanding is that while it should be unnecessary, it shouldn't
hurt to have this double wrapped `act`, but I haven't dug as deeply in
to this as the previous one. The warnings do go away when we follow the
published guidelines and let `@testing-library/user-event` handle this
on our behalf.

Signed-off-by: Noah Silas <noah@medplum.com>