Remove `lastSearch` state from `useSearch` hook
By creating a memoized object from the `resourceType` and `query` props,
we can debounce that entire object instead of debouncing a derived
`searchKey` string. That means that we can use that debounced value as
the only input to `useEffect` and see the event fire appropriately.

The main benefit here is that we no longer have an immediate `setState`
invocation inside the effect. This resolves a warning from React:

> Calling setState synchronously within an effect can trigger cascading renders

In order to memoize this correctly for query values that can be unstable
objects (example: `useSearch('Patient', { name: props.name })` would
construct a new query object every render), we retain the old behavior
of serializing the query to a string key.

Removing the condition inside the effect also makes this a bit easier to
reason about, which will help in follow up work where I intend to handle
cancellation more gracefully.

Signed-off-by: Noah Silas <noah@medplum.com>