Move test HL7 server instantiation into before/after blocks
These tests started all failing at once with this error:

> A mock server is already listening on this url

What's happening here is that I broke something else, but because this
setup and teardown was happening inside the `test` functions, the
assertion would throw and the test would fail before the teardown
happened. This caused the next test to fail when trying to start a
server.

Shifting this creation and teardown into `beforeEach`/`afterEach` blocks
lets the test runner do this cleanup even when an assertion fails,
letting the actual test errors come through more clearly.

Signed-off-by: Noah Silas <noah@medplum.com>