Fixup Agent hl7 tests for stricter MemoryRepo.createResource
This was calling `createResource` multiple times with the same ID. This
creates some ambiguity in the test (it is unclear if the existing
endpoint was updated or a new one was created).

After making `createResource` more strict to prevent this, these tests
started failing. I'm updating them to create new endpoints by explicitly
removing the `id` attribute from their input.

Signed-off-by: Noah Silas <noah@medplum.com>